Move staff only user views to new staff app
