+++ temba/orgs/tests.py
@@ -633,43 +633,6 @@ class UserTest(TembaTest):
         response = self.client.post(confirm_url, {"password": "Qwerty123"})
         self.assertRedirect(response, "/msg/")
 
-    def test_ui_permissions(self):
-        # non-logged in users can't go here
-        response = self.client.get(reverse("orgs.user_list"))
-        self.assertRedirect(response, "/users/login/")
-        response = self.client.post(reverse("orgs.user_delete", args=(self.editor.pk,)), dict(delete=True))
-        self.assertRedirect(response, "/users/login/")
-
-        # either can admins
-        self.login(self.admin)
-        response = self.client.get(reverse("orgs.user_list"))
-        self.assertRedirect(response, "/users/login/")
-        response = self.client.post(reverse("orgs.user_delete", args=(self.editor.pk,)), dict(delete=True))
-        self.assertRedirect(response, "/users/login/")
-
-        self.editor.refresh_from_db()
-        self.assertTrue(self.editor.is_active)
-
-    def test_ui_management(self):
-        # only customer support gets in on this sweet action
-        self.login(self.customer_support)
-
-        # one of our users should belong to a bunch of orgs
-        for i in range(5):
-            org = Org.objects.create(
-                name=f"Org {i}", timezone=ZoneInfo("Africa/Kigali"), created_by=self.user, modified_by=self.user
-            )
-            org.add_user(self.admin, OrgRole.ADMINISTRATOR)
-
-        response = self.client.get(reverse("orgs.user_list"))
-        self.assertEqual(200, response.status_code)
-
-        response = self.client.post(reverse("orgs.user_delete", args=(self.editor.pk,)), {"delete": True})
-        self.assertEqual(reverse("orgs.user_list"), response["Temba-Success"])
-
-        self.editor.refresh_from_db()
-        self.assertFalse(self.editor.is_active)
-
     @mock_mailroom
     def test_release(self, mr_mocks):
         token = APIToken.create(self.org, self.admin)
@@ -3301,124 +3264,6 @@ class OrgCRUDLTest(TembaTest, CRUDLTestMixin):
 
 
 class UserCRUDLTest(TembaTest, CRUDLTestMixin):
-    def test_list(self):
-        list_url = reverse("orgs.user_list")
-
-        self.assertStaffOnly(list_url)
-
-        response = self.requestView(list_url, self.customer_support)
-        self.assertEqual(8, len(response.context["object_list"]))
-        self.assertEqual("/staff/users/all", response.headers[TEMBA_MENU_SELECTION])
-
-        response = self.requestView(list_url + "?filter=beta", self.customer_support)
-        self.assertEqual(set(), set(response.context["object_list"]))
-
-        response = self.requestView(list_url + "?filter=staff", self.customer_support)
-        self.assertEqual({self.customer_support, self.superuser}, set(response.context["object_list"]))
-
-        response = self.requestView(list_url + "?search=admin@nyaruka.com", self.customer_support)
-        self.assertEqual({self.admin}, set(response.context["object_list"]))
-
-        response = self.requestView(list_url + "?search=admin@nyaruka.com", self.customer_support)
-        self.assertEqual({self.admin}, set(response.context["object_list"]))
-
-        response = self.requestView(list_url + "?search=Andy", self.customer_support)
-        self.assertEqual({self.admin}, set(response.context["object_list"]))
-
-    def test_read(self):
-        read_url = reverse("orgs.user_read", args=[self.editor.id])
-
-        # this is a customer support only view
-        self.assertStaffOnly(read_url)
-
-        response = self.requestView(read_url, self.customer_support)
-        self.assertEqual(200, response.status_code)
-
-    def test_update(self):
-        update_url = reverse("orgs.user_update", args=[self.editor.id])
-
-        # this is a customer support only view
-        self.assertStaffOnly(update_url)
-
-        response = self.requestView(update_url, self.customer_support)
-        self.assertEqual(200, response.status_code)
-
-        alphas = Group.objects.get(name="Alpha")
-        betas = Group.objects.get(name="Beta")
-        current_password = self.editor.password
-
-        # submit without new password
-        response = self.requestView(
-            update_url,
-            self.customer_support,
-            post_data={
-                "email": "eddy@nyaruka.com",
-                "first_name": "Edward",
-                "last_name": "",
-                "groups": [alphas.id, betas.id],
-            },
-        )
-        self.assertEqual(302, response.status_code)
-
-        self.editor.refresh_from_db()
-        self.assertEqual("eddy@nyaruka.com", self.editor.email)
-        self.assertEqual("eddy@nyaruka.com", self.editor.username)  # should match email
-        self.assertEqual(current_password, self.editor.password)
-        self.assertEqual("Edward", self.editor.first_name)
-        self.assertEqual("", self.editor.last_name)
-        self.assertEqual({alphas, betas}, set(self.editor.groups.all()))
-
-        # submit with new password and one less group
-        response = self.requestView(
-            update_url,
-            self.customer_support,
-            post_data={
-                "email": "eddy@nyaruka.com",
-                "new_password": "Asdf1234",
-                "first_name": "Edward",
-                "last_name": "",
-                "groups": [alphas.id],
-            },
-        )
-        self.assertEqual(302, response.status_code)
-
-        self.editor.refresh_from_db()
-        self.assertEqual("eddy@nyaruka.com", self.editor.email)
-        self.assertEqual("eddy@nyaruka.com", self.editor.username)
-        self.assertNotEqual(current_password, self.editor.password)
-        self.assertEqual("Edward", self.editor.first_name)
-        self.assertEqual("", self.editor.last_name)
-        self.assertEqual({alphas}, set(self.editor.groups.all()))
-
-    @mock_mailroom
-    def test_delete(self, mr_mocks):
-        delete_url = reverse("orgs.user_delete", args=[self.editor.id])
-
-        # this is a customer support only view
-        self.assertStaffOnly(delete_url)
-
-        response = self.requestView(delete_url, self.customer_support)
-        self.assertEqual(200, response.status_code)
-        self.assertNotContains(response, "Nyaruka")  # editor doesn't own this org
-
-        # make editor the owner of the org
-        OrgMembership.objects.filter(org=self.org, role_code=OrgRole.ADMINISTRATOR.code).delete()
-        OrgMembership.objects.filter(org=self.org, role_code=OrgRole.VIEWER.code).delete()
-        OrgMembership.objects.filter(org=self.org, role_code=OrgRole.AGENT.code).delete()
-
-        response = self.requestView(delete_url, self.customer_support)
-        self.assertEqual(200, response.status_code)
-        self.assertContains(response, "Nyaruka")
-
-        response = self.requestView(delete_url, self.customer_support, post_data={})
-        self.assertEqual(reverse("orgs.user_list"), response["Temba-Success"])
-
-        self.editor.refresh_from_db()
-        self.assertFalse(self.editor.is_active)
-
-        self.org.refresh_from_db()
-        self.assertFalse(self.org.is_active)
-
     def test_account(self):
         self.login(self.agent)
