+++ temba/staff/urls.py
@@ -1,0 +1,3 @@
+from .views import UserCRUDL
+
+urlpatterns = UserCRUDL().as_urlpatterns()
